Add custom formatter to generate factory tables
When someone runs a RSpec test with --format SlowFactoryFormatter
this formatter will generate a table of factories that take a long time
to setup.

My definition of a slow factories are factories that take more then 0.5
seconds to setup or a test where more 5 factories are created. The
definition of a slow factory is naive at the moment and hopefully the
logic around determining fast or slow factories will improve, for now
this should suffice.