Display total factory time when dumping summary
Previously the table generated and displayed to a user would only
contain the number of seconds it took to setup slow factories when a
set of tests were run. Therefore when eye balling the table it would be
hard to workout how long in proportion to the time it took to run the
tests was spent setting up factories.

This commit uses RSpec::Core SummaryNotification#duration to find
out how long the tests took to run. Then it uses that to calculate the
percentage of time spent setting up slow factories. This information
always displayed. My hope is that this will help me identify just how
long some of my tests spend setting up factories.

 #50-character subject line