Migrate test db
Previously I tried to migrate the db and by default if an env is not supplied
then development db is migrated. Which would mean that the migrations had not
been run for the test db and when my test tries to run `bundle exec rspec`
AR would raise an error and request me to run the migrations for the test db.

This commit fixes the problem by explicitly running the migrations only for the
test db.