Round seconds taken to two decimal places
Interestingly enough calling `Float#round(2)` on a floating point number with 1
decimal place returns a number floating point num with 1 decimal place.

Which is a problem in this gem as I want to return numbers to two decimal
places. This commit fixes the issue by using `String#%` to format the float to
2 decimal places.